+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-C-lb1_JK.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-DckclpLB.js";import{_ as a}from"./mermaid.core-dp5P5HSb.js";import"./chunk-5VM5RSS4-Cx3oIye-.js";import"./chunk-XXDRQBXY-CqsmFg7-.js";import"./chunk-POPQ4Y6H-CspH_N3l.js";import"./chunk-F27PBJKO-DL7yQK3D.js";import"./channel-CYtQbftK.js";import"./index-SKsDQeI6.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};